requirements.txt
connexion == 2.1.0
swagger-ui-bundle == 0.0.2
python_dateutil == 2.6.0
setuptools >= 21.0.0
gunicorn == 19.9.0
PyJWT==1.7.1
pymongo==3.7.2
Flask-Testing==0.7.1
six
